moved sdk e2e tests to separate package
