+++ packages/user-journey/README.md
@@ -1,0 +1,3 @@
+# User-journey
+
+This package is an e2e test package. It encapsulate some core usage scenarios for armory stack. In order to run, all
