+++ packages/user-journey/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "user-journey",
+  "version": "0.0.1",
+  "description": "it's a test suite that covers core scenario for narval's armory.",
+  "dependencies": {
+    "@narval-xyz/armory-sdk": "0.8.3"
+  },
+  "author": "",
+  "license": "ISC"
+}
