+++ packages/user-journey/project.json
@@ -1,0 +1,21 @@
+{
+  "name": "user-journey",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "packages/user-journey/src",
+  "projectType": "library",
+  "tags": [],
+  "targets": {
+    "lint": {
+      "executor": "@nx/eslint:lint"
+    },
+    "test:e2e": {
+      "executor": "@nx/jest:jest",
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+      "options": {
+        "jestConfig": "packages/armory-sdk/jest.e2e.ts",
+        "verbose": true,
+        "runInBand": true
+      }
+    }
+  }
+}
