+++ packages/user-journey/src/__test__/e2e/criterion/check-approvals.spec.ts
