+++ packages/user-journey/src/__test__/e2e/criterion/check-rate-limit.spec.ts
@@ -1,8 +1,5 @@
 import { Action, entitiesSchema, FIXTURE, Request } from '@narval/policy-engine-shared'
 import { v4 } from 'uuid'
-import defaultEntities from '../../resource/entity/test.default.json'
-import adminPermitAll from '../../resource/policy/checkPrincipalRole/admin-permit-all.json'
-import memberTwoTransferPerDay from '../../resource/policy/checkRateLimit/members-2-transfer-per-day.json'
 import {
   buildAuthClient,
   buildPolicy,
@@ -12,6 +9,9 @@ import {
   getAuthHost,
   saveDataStore
 } from '../../util/setup'
+import defaultEntities from '../resource/entity/test.default.json'
+import adminPermitAll from '../resource/policy/checkPrincipalRole/admin-permit-all.json'
+import memberTwoTransferPerDay from '../resource/policy/checkRateLimit/members-2-transfer-per-day.json'
 
 const systemManagerHexPk = FIXTURE.UNSAFE_PRIVATE_KEY.Root
 const ericPrivateKey = FIXTURE.UNSAFE_PRIVATE_KEY.Eric
