+++ packages/user-journey/src/__test__/e2e/criterion/check-spending-limit.spec.ts
