+++ packages/user-journey/src/__test__/e2e/scenarii/user-journeys.spec.ts
