+++ packages/user-journey/src/resource/entity/defi-interaction.json
