+++ packages/user-journey/src/resource/policy/checkPrincipalRole/admin-permit-all.json
